feat: custom hook for size windows
